[disboardreminder] this cog is scuffed, i hope to

+++ disboardreminder/disboardreminder.py
@@ -209,7 +209,7 @@ class DisboardReminder(commands.Cog):
             return
         clean = data["clean"]
 
-        if clean and message.author != message.guild.me and message.author.id != 302050872383242240 and message.channel == clean:
+        if clean and message.author != message.guild.me and message.author.id != 302050872383242240 and message.channel == bumpChannel:
             if message.channel.permissions_for(message.guild.me).manage_messages:
                 await asyncio.sleep(5)
                 try:
@@ -234,7 +234,7 @@ class DisboardReminder(commands.Cog):
 
             await self.bump_timer(message.guild, 7200)
         else:
-            if message.channel.permissions_for(message.guild.me).manage_messages and clean:
+            if message.channel.permissions_for(message.guild.me).manage_messages and clean and message.channel == bumpChannel:
                 await asyncio.sleep(5)
                 try:
                     await message.delete()
